product discreption and the search added

+++ src/java/connections/DBHandler.java
@@ -43,7 +43,7 @@ public interface DBHandler {
 
     public ArrayList<Product> searchProduct(String query);
 
-    public ArrayList<Product>searchProductByPrice(int price);
+    public ArrayList<Product>searchProductByPrice(double price);
     
     //shopping cart
